Create czechia
Open-access data relevant to the Czech Republic, available to test 'archaeoModelling', a QGIS plugin.  Links are provided so additional geospatial data, elevation and hydrology, can be downloaded if the user chooses;  The downloads are 7.2 G and 1.3 G respectively.  Although the documentation contains ideas for coping with very large files, it is not a GIS learning resource.