Rousay_data
Rousay, a small island off the North coast of mainland Scotland, is known as 'the Egypt of the North' because it is so rich in archaeological treasures.  This directory of testing data is all open-access, and was either processed using QGIS tools or a QGIS processing tool written to clean a specific data.  That tool is included and is open-source